requirements.txt
requests
bitso-py